requirements.txt
fpdf
numpy
opencv-python
Pillow
psutil
PyQt51
tensorflow
